Allow 'view-users' permission to see the user credentials tab
Customer Service needs access to the "credentials" tab for users in Admin Console. Currently only users with manage-users permissions have this access. This story calls for Admin Console users with "view-users" permission to see the user credentials tab.This is a back port of the issue #17174 in Community Keycloak project.

Closes CIAM-5110